requirements.txt
pytest>=7.1.2
greenlet==2.0.1
pip==22.3.1
psycopg2==2.9.5
python-dotenv==0.21.0
setuptools==65.6.3
SQLAlchemy==1.4.44
wheel==0.38.4
pandas==1.5.1 